ClimaShallowWater.jl

ClimaShallowWater.jl is an implementation of the shallow water equations using ClimaCore.jl.

Test cases

ClimaShallowWater.SphericalParameters — Type

SphericalParameters

Physical parameters needed for all spherical simulations

Fields

R: Radius of earth

Ω: Rotation rate of earth

g: Gravitational constant

ν: Hyperdiffusion coefficient

α: angle between the north pole and the center of the top cube panel

ClimaShallowWater.SteadyStateTest — Type

SteadyStateTest()

Test case 2 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

This test case gives the steady-state solution to the non-linear shallow water equations. It consists of solid body rotation or zonal flow with the corresponding geostrophic height field. This can be run with an angle α that represents the angle between the north pole and the center of the top cube panel.

Fields

params: Physical parameters

u0: advection velocity

h0: peak of analytic height field

ClimaShallowWater.SteadyStateCompactTest — Type

SteadyStateCompactTest()

Test case 3 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

This test case gives the steady-state solution to the non-linear shallow water equations with nonlinear zonal geostrophic flow with compact support.

Fields

params: Physical parameters

u0: advection velocity

h0: peak of analytic height field

ϕᵦ: latitude lower bound for coordinate transformation parameter

ϕₑ: latitude upper bound for coordinate transformation parameter

xₑ: velocity perturbation parameter

ClimaShallowWater.MountainTest — Type

MountainTest()

Test case 5 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

It represents a zonal flow over an isolated mountain, where the governing equations describe a global steady-state nonlinear zonal geostrophic flow, with a corresponding geostrophic height field over a non-uniform reference surface h_s.

Fields

params: Physical parameters

u0: advection velocity

h0: peak of analytic height field

a: radius of conical mountain

λc: center of mountain long coord, shifted by 180 compared to the paper, because our $λ ∈ [-180, 180]$ (in the paper it was 270, with $λ ∈ [0, 360]$)

ϕc: latitude coordinate for center of mountain

h_s0: mountain peak height

ClimaShallowWater.RossbyHaurwitzTest — Type

RossbyHaurwitzTest <: AbstractSphereTestCase

Test case 6 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

It represents the solution of the nonlinear barotropic vorticity equation on the sphere

Fields

params: Physical parameters

a: velocity amplitude parameter

h0: peak of analytic height field

ω: vorticity amplitude parameter (1/sec)

K: vorticity amplitude parameter (1/sec)

ClimaShallowWater.BarotropicInstabilityTest — Type

BarotropicInstabilityTest{FT, P} <: AbstractSphereTestCase

Test case from Joseph Galewsky, Richard K. Scott, Lorenzo M. Polvani (2004) (and also 7.6 of Paul A. Ullrich, Christiane Jablonowski, Bram van Leer (2010)).

This test case consists of a zonal jet with compact support at a latitude of 45°. A small height disturbance is then added, which causes the jet to become unstable and collapse into a highly vortical structure.

Fields

params: Physical parameters

u_max: maximum zonal velocity

αₚ: mountain shape parameters

βₚ: mountain shape parameters

h0: peak of balanced height field from Tempest https://github.com/paullric/tempestmodel/blob/master/test/shallowwater_sphere/BarotropicInstabilityTest.cpp#L86

h_hat: local perturbation peak height

ϕ₀: southern jet boundary

ϕ₁: northern jet boundary

ϕ₂: height perturbation peak location

eₙ: zonal velocity decay parameter

References

Galewsky2004

Joseph Galewsky, Richard K. Scott, Lorenzo M. Polvani, An initial-value problem for testing numerical models of the global shallow-water equations, Tellus A: Dynamic Meteorology and Oceanography, 56(5), 429-440, 2004.

Ullrich2010

Paul A. Ullrich, Christiane Jablonowski, Bram van Leer, High-Order Finite-Volume Methods for the Shallow-Water Equations on the Sphere, J. Comput. Phys., 229(17), 6104–6134, 2010.

Williamson1992

David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber, A standard test set for numerical approximations to the shallow water equations in spherical geometry, Journal of Computational Physics, 102(1), 211-224, 1992.