API

ClimaShallowWater.initial_condition — Function

initial_condition(space, test::AbstractSphereTestCase)

Construct the initial state for the test case test on the space provided.

ClimaShallowWater.surface_height_field — Function

surface_height_field(space, test::AbstractSphereTestCase)

Construct a field on space of the surface topography height h_s for the test case test.

Set to be a field of zeros if not otherwise defined.

ClimaShallowWater.coriolis_field — Function

coriolis_field(space, test::AbstractSphereTestCase)

Construct a field on space of the angular coriolis velocity for the test case test.

By default it is

\[f = 2Ω (\sin(\phi) \cos(\alpha) - \cos(\lambda) \cos(\phi) \sin(\alpha)\]

where $\alpha$ is the angle between the north pole and the center of the top, and $\lambda$ and $\phi$ are the longitude and latitude of the grid point.