Test cases

ClimaShallowWater.SphericalParameters — Type

SphericalParameters

Physical parameters needed for all spherical simulations

Fields

planet_radius: Radius of earth

rotation_rate: Rotation rate of earth

grav: Gravitational constant

hyperdiff_coefficient: Hyperdiffusion coefficient

angle_α: angle between the north pole and the center of the top cube panel

ClimaShallowWater.SteadyStateTest — Type

SteadyStateTest()

Test case 2 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

This test case gives the steady-state solution to the non-linear shallow water equations. It consists of solid body rotation or zonal flow with the corresponding geostrophic height field. This can be run with an angle α that represents the angle between the north pole and the center of the top cube panel.

Fields

common_parameters: Physical parameters

advection_velocity: advection velocity

peak_analytic_height_field_parameter: constant for computing peak of analytic height field

peak_analytic_height_field: peak of analytic height field

ClimaShallowWater.SteadyStateCompactTest — Type

SteadyStateCompactTest()

Test case 3 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

This test case gives the steady-state solution to the non-linear shallow water equations with nonlinear zonal geostrophic flow with compact support.

Fields

common_parameters: Physical parameters

coord_transform_lower_bound: latitude lower bound for coordinate transformation parameter

coord_transform_upper_bound: latitude upper bound for coordinate transformation parameter

velocity_perturbation: velocity perturbation parameter

advection_velocity: advection velocity

peak_analytic_height_field_parameter: constant for computing peak of analytic height field

peak_analytic_height_field: peak of analytic height field

ClimaShallowWater.MountainTest — Type

MountainTest()

Test case 5 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

It represents a zonal flow over an isolated mountain, where the governing equations describe a global steady-state nonlinear zonal geostrophic flow, with a corresponding geostrophic height field over a non-uniform reference surface h_s.

Fields

common_parameters: Physical parameters

advection_velocity: advection velocity

peak_analytic_height_field: peak of analytic height field

velocity_amplitude: radius of conical mountain

mountain_center_longitude: center of mountain long coord, shifted by 180 compared to the paper, because our $λ ∈ [-180, 180]$ (in the paper it was 270, with $λ ∈ [0, 360]$)

mountain_center_latitude: latitude coordinate for center of mountain

mountain_peak_height: mountain peak height

ClimaShallowWater.RossbyHaurwitzTest — Type

RossbyHaurwitzTest <: AbstractSphereTestCase

Test case 6 of David L. Williamson, John B. Drake, James J. Hack, Rüdiger Jakob, Paul N. Swarztrauber (1992).

It represents the solution of the nonlinear barotropic vorticity equation on the sphere

Fields

common_parameters: Physical parameters

velocity_amplitude: velocity amplitude parameter

peak_analytic_height_field: peak of analytic height field

vorticity_amplitude_ω: vorticity amplitude parameter (1/sec)

vorticity_amplitude_K: vorticity amplitude parameter (1/sec)

ClimaShallowWater.BarotropicInstabilityTest — Type

BarotropicInstabilityTest{FT} <: AbstractSphereTestCase

Test case from Joseph Galewsky, Richard K. Scott, Lorenzo M. Polvani (2004) (and also 7.6 of Paul A. Ullrich, Christiane Jablonowski, Bram van Leer (2010)).

This test case consists of a zonal jet with compact support at a latitude of 45°. A small height disturbance is then added, which causes the jet to become unstable and collapse into a highly vortical structure.

Fields

common_parameters: Physical parameters

max_zonal_velocity: maximum zonal velocity

mountain_shape_α: mountain shape parameters

mountain_shape_β: mountain shape parameters

peak_analytic_height_field: peak of balanced height field from Tempest https://github.com/paullric/tempestmodel/blob/master/test/shallowwater_sphere/BarotropicInstabilityTest.cpp#L86

local_perturb_peak_height: local perturbation peak height

southern_jet_boundary: southern jet boundary

northern_jet_boundary: northern jet boundary

height_perturbation_peak_location: height perturbation peak location